Switch out Cucumber for Spinach
Spinach seems like a much nicer solution - steps are well defined and in
obvious places. This is mostly a straight conversion, although I have reworked
some of the tests - the main change being 'am on' -> 'visit' for URLs.

While doing this, I noticed a few features that have multiple steps where not
really needed - a 'submit form with data' rather than 'fill in form' and
'submit' would be much easier to read (and less code)