Return bad request on certain accepted answer calls
If the accepted_answer_id is neither present in the request parameters or on
the question, return a bad request. This shouldn't be a scenario encountered in
everyday browsing, but was an edge case I discoverd while considering
refactoring.

This also adds a spec to ensure the accepted_answer_id is removed when it is
set on the question but nil in the request