Store an Authorization as soon as the Omniauth callback is called
Previously the provider/email/uid fields of the callback were 'stored'
in hidden fields in the user create form. We now create an
Authorization instance immediately, and keep a reference to that ID
in the session.

A before_action callback will tidy up this stored Authorization if the
user clicks away from the 'create user' page before creating a user, to
avoid filling the database with useless Authorizations - and allowing
the user to sign up again in the future.